django/requirement.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-js-asset==1.2.2
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
python-decouple==3.3
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
